Update possibly outdated public url example
I know it is mentioned as an example public URL, but I found my public url was different, and because I was getting a 403 error when trying to grab my image via the example url, it made me believe something was up with my permissions. Does it need to be changed?